Add files via upload
Made use of different analytical graphs such as bar charts and pie chart to depict the characteristic of the dataset. Also, heatmaps were used to demonstrate the correlation between multiple features of this dataset. EDA is a stage designed for analysing data, which helps us with figuring out the underlying patterns of our data and hence choose a more suitable technique or model to apply on our dataset to get the best possible result.